requirements/base.txt
# Wheel 0.25+ needed to install certain packages on CPython 3.5+
# like Pillow and psycopg2
# See http://bitly.com/wheel-building-fails-CPython-35
# Verified bug on Python 3.5.1
wheel==0.29.0

# Bleeding edge Django
django==1.10.7

# Configuration
django-environ==0.4.1



# Forms
django-braces==1.10.0
django-crispy-forms==1.6.1

# Models
django-model-utils==2.6

# Images
Pillow==3.4.2

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.31.0


# Python-PostgreSQL Database Adapter
psycopg2==2.7.3.2

# Unicode slugification
django-autoslug==1.9.3
awesome-slugify==1.6.5

# Time zones support
pytz==2016.7

# Redis support
django-redis==4.6.0
redis>=2.10.5


celery==3.1.24




# Your custom requirements go here
djangorestframework==3.5.3
django-imagekit==3.3
django-multiselectfield==0.1.4
django-haystack==2.5.1
elasticsearch==5.0.1
django-filter==1.0.2
geocoder==1.19.0
django-import-export==1.0.0
django-rest-auth==0.9.1
django-nested-admin==3.0.17
django-froala-editor==2.5.1
django-webpack-loader==0.5.0
